+++ routes/index.js
@@ -16,11 +16,8 @@ router.get('/', async (req, res) => {
         });
 
     } catch (err) {
-        console.log(err.message);
         books = [];
     }
-    console.log(req.user);
-    console.log(req.user.ID, req.user.email);
     res.render('index', {
         books
     });
