+++ utils/middleware.js
@@ -1,0 +1,15 @@
+function checkAuthentication(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/authors/login');
+}
+
+function checkNotAuthentication(req, res, next) {
+    if (!req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/');
+}
+
+module.exports = {checkAuthentication, checkNotAuthentication};
