+++ views/authors/_form_inputs.ejs
@@ -11,7 +11,7 @@
 <div class="form-row">
     <div class="form-item">
         <label for="password">Password</label>
-        <input  type="password" name="password" id="password" value="<%= author?.password %>" required>
+        <input  type="password" name="password" id="password" value="" required>
     </div>
 
     <div class="form-item">
