+++ views/layouts/layout.ejs
@@ -22,7 +22,9 @@
 <body>
     <div class="container">
         <%- include('../partials/header.ejs') %>
+
         <div class="content">
+            <%- include('../partials/showMessages.ejs') %>
             <%- include('../partials/errorMessage.ejs') %>
             <%- body %>
         </div>
