+++ views/partials/showMessages.ejs
@@ -1,0 +1,11 @@
+<% if (messages.info) { %>
+    <div class="mes info"><%= messages.info %></div>
+<% } %>
+
+<% if (messages.error) { %>
+    <div class="mes error"><%= messages.error %></div>
+<% } %>
+
+<% if (messages.success) { %>
+    <div class="mes success"><%= messages.success %></div>
+<% } %>
